+++ src/popover.ts
@@ -1,66 +1,4 @@
-const popovers = new Set<HTMLElement>();
-
-const handleChildListMutation = (mutation: MutationRecord) => {
-  if (mutation.addedNodes.length > 0) {
-    mutation.addedNodes.forEach((node) => {
-      if (node instanceof HTMLElement) {
-        if (node.hasAttribute('popover')) {
-          popovers.add(node);
-        }
-      }
-    });
-  }
-
-  mutation.removedNodes.forEach((node) => {
-    if (node instanceof HTMLElement && node.hasAttribute('popover')) {
-      popovers.delete(node);
-    }
-  });
-};
-
-const handlePopoverAttributeMutation = (mutation: MutationRecord) => {
-  if (mutation.target instanceof HTMLElement) {
-    if (mutation.target.hasAttribute('popover')) {
-      popovers.add(mutation.target);
-    } else {
-      popovers.delete(mutation.target);
-    }
-  }
-};
-
-const handleMutation = (mutationList: MutationRecord[]) => {
-  mutationList.forEach((mutation) => {
-    switch (mutation.type) {
-      case 'attributes':
-        switch (mutation.attributeName) {
-          case 'popover':
-            handlePopoverAttributeMutation(mutation);
-            break;
-        }
-        break;
-      case 'childList':
-        handleChildListMutation(mutation);
-        break;
-    }
-  });
-};
-
-function patchAttachShadow() {
-  Element.prototype._originalAttachShadow = Element.prototype.attachShadow;
-  Element.prototype.attachShadow = function (init) {
-    const shadowRoot = this._originalAttachShadow(init);
-
-    const observer = new MutationObserver(handleMutation);
-
-    observer.observe(shadowRoot, {
-      attributeFilter: ['popover'],
-      childList: true,
-      subtree: true,
-    });
-
-    return shadowRoot;
-  };
-}
+import { observeDocumentOrShadowRootMutations } from './mutation';
 
 export function isSupported() {
   return (
@@ -73,58 +11,18 @@ export function isSupported() {
 const notSupportedMessage =
   'Not supported on element that does not have valid popover attribute';
 
-// class PopoverObserver {
-//   #popovers = new Set<HTMLElement>();
-
-//   #observer = new MutationObserver(this.#callback);
-
-//   #userPopoverChanged(target: Node) {
-//     console.log(target);
-//   }
-
-//   #callback(mutationList: MutationRecord[]) {
-//     mutationList.forEach((mutation) => {
-//       switch (mutation.type) {
-//         case 'attributes':
-//           switch (mutation.attributeName) {
-//             case 'popover':
-//               this.#userPopoverChanged(mutation.target);
-//               break;
-//           }
-//           break;
-//       }
-//     });
-//   }
-
-//   add(tree: Document | ShadowRoot) {
-//     const popovers = tree.querySelectorAll('[popover]');
-//     console.log(tree);
-
-//     popovers.forEach((popover) => {
-//       console.log(popover instanceof HTMLElement);
-//       if (popover instanceof HTMLElement) {
-//         this.#popovers.add(popover);
-//         console.log([...this.#popovers]);
-//       }
-//     });
-
-//     this.#observer.observe(tree, {
-//       attributeFilter: ['popover'],
-//       childList: true,
-//       subtree: true,
-//     });
-//   }
-
-//   // get popovers() {
-//   //   return this.#popovers;
-//   // }
-// }
+function patchAttachShadow(callback: (shadowRoot: ShadowRoot) => void) {
+  Element.prototype._originalAttachShadow = Element.prototype.attachShadow;
+  Element.prototype.attachShadow = function (init) {
+    const shadowRoot = this._originalAttachShadow(init);
+    callback(shadowRoot);
+    return shadowRoot;
+  };
+}
 
 export function apply() {
-  // const popoverObserver = new PopoverObserver();
-  // popoverObserver.add(document);
-  // patchAttachShadow((shadowRoot) => popoverObserver.add(shadowRoot));
-  patchAttachShadow();
+  observeDocumentOrShadowRootMutations(document);
+  patchAttachShadow(observeDocumentOrShadowRootMutations);
 
   const visibleElements = new WeakSet<HTMLElement>();
